+++ R/voson.tcn-package.R
@@ -3,7 +3,6 @@
 #' @docType package
 #' @keywords internal
 "_PACKAGE"
-#' @importFrom magrittr "%>%" "%<>%"
 #' @importFrom rlang ".data"
 #' @importFrom stats na.omit
 NULL